Split out SentenceHelper
